WoT restructure: rename utils/cellCoords.ts -> utils/hitTest.ts (avoid clash with cell/coords CellCoords)
