+++ handsontable/src/3rdparty/walkontable/src/utils/hitTest.ts
